Solving some mapper details

+++ app/src/main/java/co/bold/weather/data/model/Day.kt
@@ -6,7 +6,7 @@ data class Day(
     @SerializedName("maxtemp_c") var maxtempC: Double? = null,
     @SerializedName("maxtemp_f") var maxtempF: Double? = null,
     @SerializedName("mintemp_c") var mintempC: Double? = null,
-    @SerializedName("mintemp_f") var mintempF: Int? = null,
+    @SerializedName("mintemp_f") var mintempF: Double? = null,
     @SerializedName("avgtemp_c") var avgtempC: Double? = null,
     @SerializedName("avgtemp_f") var avgtempF: Double? = null,
     @SerializedName("maxwind_mph") var maxwindMph: Double? = null,
@@ -15,12 +15,12 @@ data class Day(
     @SerializedName("totalprecip_in") var totalprecipIn: Double? = null,
     @SerializedName("totalsnow_cm") var totalsnowCm: Int? = null,
     @SerializedName("avgvis_km") var avgvisKm: Double? = null,
-    @SerializedName("avgvis_miles") var avgvisMiles: Int? = null,
-    @SerializedName("avghumidity") var avghumidity: Int? = null,
-    @SerializedName("daily_will_it_rain") var dailyWillItRain: Int? = null,
-    @SerializedName("daily_chance_of_rain") var dailyChanceOfRain: Int? = null,
-    @SerializedName("daily_will_it_snow") var dailyWillItSnow: Int? = null,
-    @SerializedName("daily_chance_of_snow") var dailyChanceOfSnow: Int? = null,
+    @SerializedName("avgvis_miles") var avgvisMiles: Double? = null,
+    @SerializedName("avghumidity") var avghumidity: Double? = null,
+    @SerializedName("daily_will_it_rain") var dailyWillItRain: Double? = null,
+    @SerializedName("daily_chance_of_rain") var dailyChanceOfRain: Double? = null,
+    @SerializedName("daily_will_it_snow") var dailyWillItSnow: Double? = null,
+    @SerializedName("daily_chance_of_snow") var dailyChanceOfSnow: Double? = null,
     @SerializedName("condition") var condition: Condition? = Condition(),
-    @SerializedName("uv") var uv: Int? = null
+    @SerializedName("uv") var uv: Double? = null
 )
